+++ src/main/java/wooteco/subway/ui/service/PathService.java
@@ -9,7 +9,7 @@ import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.fare.Fare;
 import wooteco.subway.domain.line.Line;
-import wooteco.subway.domain.path.PathCalculator;
+import wooteco.subway.domain.path.DijkstraPathCalculator;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.request.PathRequest;
 import wooteco.subway.dto.response.PathResponse;
@@ -34,15 +34,15 @@ public class PathService {
 
     private PathResponse getPath(long source, long target, int age) {
         List<Line> lines = lineDao.findAll();
-        PathCalculator pathCalculator = PathCalculator.from(lines);
+        DijkstraPathCalculator dijkstraPathCalculator = DijkstraPathCalculator.from(lines);
 
         Station sourceStation = stationDao.findById(source)
                 .orElseThrow(() -> new IllegalArgumentException("조회하고자 하는 상행역이 존재하지 않습니다."));
         Station targetStation = stationDao.findById(target)
                 .orElseThrow(() -> new IllegalArgumentException("조회하고자 하는 하행역이 존재하지 않습니다."));
 
-        List<Station> path = pathCalculator.calculateShortestPath(sourceStation, targetStation);
-        List<Long> lineIds = pathCalculator.calculateShortestPathLines(sourceStation, targetStation);
+        List<Station> path = dijkstraPathCalculator.calculateShortestPath(sourceStation, targetStation);
+        List<Long> lineIds = dijkstraPathCalculator.calculateShortestPathLines(sourceStation, targetStation);
         int maxExtraFare = lineIds.stream()
                 .map(lineDao::findById)
                 .filter(Optional::isPresent)
@@ -50,7 +50,7 @@ public class PathService {
                 .mapToInt(Line::getExtraFare)
                 .max()
                 .orElse(0);
-        double distance = pathCalculator.calculateShortestDistance(sourceStation, targetStation);
+        double distance = dijkstraPathCalculator.calculateShortestDistance(sourceStation, targetStation);
 
         return new PathResponse(StationResponse.of(path), distance, Fare.of(distance, maxExtraFare, age));
     }
